move beacon node folder to parent folder
