+++ src/renderer/containers/BeaconNode/BeaconNode.scss
