+++ src/renderer/containers/BeaconNode/BeaconNode.stories.tsx
