+++ src/renderer/containers/BeaconNode/BeaconNodeResponseErrorPieChart.tsx
