+++ src/renderer/containers/ValidatorDetails/ValidatorDetails.scss
@@ -1,5 +1,5 @@
 @import "./ValidatorStats/ValidatorStats";
-@import "./BeaconNode/BeaconNode";
+@import "../BeaconNode/BeaconNode";
 
 .validator-details-container {
   align-items: flex-end;
